refactor: keep entry.client.js byte-identical when mount_target is unset
Split the embed entry into a separate template and only swap it in when
mount_target is set, so non-embed builds produce the exact same
entry.client.js as before this feature landed.
