+++ packages/reflex-base/src/reflex_base/constants/compiler.py
@@ -114,6 +114,9 @@ class Embed(SimpleNamespace):
     # In dev, Vite serves the source file at this URL; in prod, a shim emitted
     # by ``_emit_stable_entry_bootloader`` re-exports the hashed Vite entry.
     ENTRY_PATH = "app/entry.client.js"
+    # Embed-aware variant of the entry, swapped in only when ``mount_target``
+    # is set so non-embed builds remain byte-identical to the framework default.
+    ENTRY_EMBED_TEMPLATE = "app/entry.client.embed.js"
     # Compile-time-emitted route manifest the embed entry imports at runtime.
     MANIFEST_FILE = "__reflex_embed_manifest.js"
